Fixes #30897
Move the Generic type parameter naming conventions to the general naming conventions.
